# contracts/openapi/riverhog.v1.yaml
openapi: 3.1.0
info:
  title: RIVERHOG API
  version: 1.0.0
  summary: MVP API for hot/cold archival storage
  description: Authoritative archive resources are durable across service restarts unless a specific contract explicitly expires or releases them.
servers:
  - url: /v1
paths:
  /collection-uploads:
    post:
      summary: Create or resume one collection upload session
      description: Creates or resumes a collection upload session from a human-readable slug and expected logical file set. Riverhog normalizes the slug, mints the canonical collection id with either the server UTC upload timestamp or an optional caller-provided UTC timestamp, and resumes an existing non-finalized session when the normalized slug and manifest match.
      operationId: createOrResumeCollectionUpload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionUploadRequest"
      responses:
        "200":
          description: Collection upload session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionUploadSession"
        default:
          $ref: "#/components/responses/Error"
  /collection-upload-sessions:
    post:
      summary: Create or resume one incremental collection upload session
      description: Creates or resumes an open collection upload session from a human-readable slug without requiring a complete file manifest up front. Files are registered one at a time, uploaded through the normal collection-file upload resource, then frozen by an explicit complete call.
      operationId: createOrResumeCollectionUploadSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrResumeCollectionUploadSessionRequest"
      responses:
        "200":
          description: Collection upload session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionUploadSession"
        default:
          $ref: "#/components/responses/Error"
...
  /files:
    get:
      summary: Query files matching a target selector
      description: Returns one page of files whose projected path matches the given target selector with their full state. Returns an empty list when no files match; never 404.
      operationId: queryFiles
      parameters:
        - in: query
          name: target
          required: true
          schema: { type: string }
          description: Canonical projected-path selector. File selectors return exactly matching files; directory selectors (ending with `/`) return all files under that prefix.
        - in: query
          name: page
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: per_page
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
      responses:
        "200":
          description: Files matching the target selector
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilesResponse"
        default:
          $ref: "#/components/responses/Error"
...
components:
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    Target:
      type: string
      description: Canonical projected-path selector relative to the hot-view root. Directory selectors end with `/` and file selectors do not.
...
    FetchEntryUploadState:
      type: string
      enum: [pending, partial, byte_complete, uploaded]
...
    CoverageState:
      type: string
      enum: [none, partial, full]
    Archive:
      type: object
      required:
        [state, object_path, stored_bytes, backend, storage_class, last_uploaded_at, last_verified_at, failure]
      properties:
        state:
          type: string
          enum: [pending, uploading, uploaded, retrying, failed]
        object_path:
          oneOf:
            - type: string
            - type: "null"
        stored_bytes:
          oneOf:
            - type: integer
            - type: "null"
        backend:
          oneOf:
            - type: string
            - type: "null"
        storage_class:
          oneOf:
            - type: string
            - type: "null"
        last_uploaded_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        last_verified_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        failure:
          oneOf:
            - type: string
            - type: "null"
    CollectionManifestProofState:
      type: string
      enum: [pending, uploaded, failed]
    CollectionManifest:
      type: object
      required: [object_path, sha256, ots_object_path, ots_state]
      properties:
        object_path:
          oneOf:
            - type: string
            - type: "null"
        sha256:
          oneOf:
            - type: string
              pattern: "^[0-9a-f]{64}$"
            - type: "null"
        ots_object_path:
          oneOf:
            - type: string
            - type: "null"
        ots_state:
          $ref: "#/components/schemas/CollectionManifestProofState"
...
    CollectionSummary:
      type: object
      required:
        [
          id,
          files,
          bytes,
          hot_bytes,
          archive,
          collection_manifest,
          archive_format,
          compression,
          disc_coverage,
          disc_redundancy,
          image_coverage,
        ]
      properties:
        id:
          type: string
          description: Canonical collection id. May contain `/` and is prefix-free with every other collection id.
        files: { type: integer }
        bytes: { type: integer }
        hot_bytes: { type: integer }
        archive:
          $ref: "#/components/schemas/Archive"
        collection_manifest:
          $ref: "#/components/schemas/CollectionManifest"
        archive_format:
          type: string
        compression:
          type: string
        disc_coverage:
          $ref: "#/components/schemas/Coverage"
        disc_redundancy:
          $ref: "#/components/schemas/Coverage"
        image_coverage:
          type: array
          items:
            $ref: "#/components/schemas/CollectionCoverageImage"
...
    Coverage:
      type: object
      required: [state, bytes]
      properties:
        state:
          $ref: "#/components/schemas/CoverageState"
        bytes:
          type: integer
          minimum: 0
...
    CollectionUploadFile:
      type: object
      required: [path, bytes, sha256]
      properties:
        path:
          type: string
          description: Relative logical file path within the collection source tree.
        bytes: { type: integer }
        sha256:
          type: string
          description: Logical-file SHA-256 digest for final verification.
    CollectionUploadNotify:
      type: object
      properties:
        enabled:
          type: boolean
          description: Whether Riverhog should deliver collection lifecycle notifications for this upload.
        recipients:
          type: array
          items:
            type: string
          description: Deployment-owned recipient ids resolved through Riverhog notification config.
    CollectionUploadRequest:
      type: object
      required: [slug, files]
      properties:
        slug:
          type: string
          description: Human-readable collection slug. Riverhog normalizes this and mints the canonical collection id.
        upload_timestamp:
          oneOf:
            - type: string
            - type: "null"
          description: Optional UTC basic timestamp in `YYYYMMDDTHHMMSSZ` form to use in the canonical collection id for migration of older archival sets. Slug remains required.
        ingest_source:
          oneOf:
            - type: string
            - type: "null"
          description: Descriptive source hint from the client. Never part of collection identity.
        notify:
          oneOf:
            - $ref: "#/components/schemas/CollectionUploadNotify"
            - type: "null"
          description: Optional collection lifecycle notification routing metadata.
        files:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CollectionUploadFile"
    CreateOrResumeCollectionUploadSessionRequest:
      type: object
      required: [slug]
      properties:
        slug:
          type: string
          description: Human-readable collection slug. Riverhog normalizes this and mints or resumes the canonical collection id.
        upload_timestamp:
          oneOf:
            - type: string
            - type: "null"
          description: Optional UTC basic timestamp in `YYYYMMDDTHHMMSSZ` form to use in the canonical collection id for migration of older archival sets. Slug remains required.
        ingest_source:
          oneOf:
            - type: string
            - type: "null"
          description: Descriptive source hint from the client. Never part of collection identity.
        notify:
          oneOf:
            - $ref: "#/components/schemas/CollectionUploadNotify"
            - type: "null"
          description: Optional collection lifecycle notification routing metadata.
...
    CollectionUploadFileState:
      type: object
      required: [path, bytes, sha256, upload_state, uploaded_bytes, upload_state_expires_at]
      properties:
        path: { type: string }
        bytes: { type: integer }
        sha256: { type: string }
        upload_state:
          $ref: "#/components/schemas/FetchEntryUploadState"
        uploaded_bytes:
          type: integer
          description: Accepted upload bytes for this logical file's plaintext stream.
        upload_state_expires_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
...
    CollectionUploadSession:
      type: object
      required:
        [
          collection_id,
          ingest_source,
          state,
          files_total,
          files_pending,
          files_partial,
          files_uploaded,
          hot_promoted_files,
          bytes_total,
          uploaded_bytes,
          hot_promoted_bytes,
          missing_bytes,
          upload_state_expires_at,
          files,
          collection,
        ]
      properties:
        collection_id:
          type: string
          description: Server-minted canonical collection id for this upload session.
        ingest_source:
          oneOf:
            - type: string
            - type: "null"
        state:
          type: string
          enum: [open, uploading, archiving, finalized, failed, canceled, expired]
        files_total: { type: integer }
        files_pending: { type: integer }
        files_partial: { type: integer }
        files_uploaded: { type: integer }
        hot_promoted_files:
          type: integer
          description: Files verified in the hot collection namespace during finalization.
        bytes_total: { type: integer }
        uploaded_bytes: { type: integer }
        hot_promoted_bytes:
          type: integer
          description: Bytes verified in the hot collection namespace during finalization.
        missing_bytes: { type: integer }
        upload_state_expires_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        latest_failure:
          oneOf:
            - type: string
            - type: "null"
          description: Latest collection archive failure, when archiving exhausted retries.
        archive_phase:
          oneOf:
            - type: string
            - type: "null"
          description: Current long-running archive phase for upload finalization.
        archive_phase_updated_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        archive_object_path:
          oneOf:
            - type: string
            - type: "null"
        archive_uploaded_bytes:
          oneOf:
            - type: integer
            - type: "null"
        archive_total_bytes:
          oneOf:
            - type: integer
            - type: "null"
        archive_uploaded_parts:
          oneOf:
            - type: integer
            - type: "null"
        archive_total_parts:
          oneOf:
            - type: integer
            - type: "null"
        notify:
          oneOf:
            - $ref: "#/components/schemas/CollectionUploadNotify"
            - type: "null"
        files:
          type: array
          items:
            $ref: "#/components/schemas/CollectionUploadFileState"
        collection:
          oneOf:
            - $ref: "#/components/schemas/CollectionSummary"
            - type: "null"
...
    CollectionCoverageImage:
      type: object
      required:
        [
          id,
          filename,
          disc_redundancy_state,
          discs_required,
          discs_registered,
          discs_verified,
          discs_missing,
          covered_paths,
          discs,
        ]
      properties:
        id:
          type: string
          pattern: "^[0-9]{8}T[0-9]{6}Z$"
        filename:
          type: string
        disc_redundancy_state:
          $ref: "#/components/schemas/CoverageState"
        discs_required:
          type: integer
          minimum: 1
        discs_registered:
          type: integer
          minimum: 0
        discs_verified:
          type: integer
          minimum: 0
        discs_missing:
          type: integer
          minimum: 0
        covered_paths:
          type: array
          items:
            type: string
          description: Relative collection paths within this collection that the finalized image represents.
        covered_paths_total:
          type: integer
          minimum: 0
          description: Total covered path count before any `coverage_path_limit` truncation.
        discs:
          type: array
          items:
            $ref: "#/components/schemas/DiscSummary"
    DiscSummary:
      type: object
      required:
        [
          disc_id,
          label_text,
          location,
          created_at,
          state,
          verification_state,
          history,
          image_id,
        ]
      properties:
        disc_id:
          type: string
          description: Generated disc id, unique within one finalized image.
        label_text:
          type: string
          description: Exact operator-facing disc label text for this generated disc slot.
        location:
          type: string
          nullable: true
          description: Mutable physical location label. Not part of disc identity.
        created_at: { type: string, format: date-time }
        state:
          type: string
          enum: [needed, burning, verified, registered, lost, damaged, retired]
        verification_state:
          type: string
          enum: [pending, verified, failed]
        history:
          type: array
          items:
            $ref: "#/components/schemas/DiscHistory"
        image_id:
          type: string
          pattern: "^[0-9]{8}T[0-9]{6}Z$"
          description: Finalized image id for this disc.
        filename:
          type: string
          nullable: true
          description: Stable ISO filename for the finalized image artifact.
...
    DiscHistory:
      type: object
      required: [at, event, state, verification_state, location]
      properties:
        at: { type: string, format: date-time }
        event: { type: string }
        state:
          type: string
          enum: [needed, burning, verified, registered, lost, damaged, retired]
        verification_state:
          type: string
          enum: [pending, verified, failed]
        location:
          type: string
          nullable: true
...
    FileRecord:
      type: object
      required: [target, collection, path, bytes, sha256, hot, disc_coverage]
      properties:
        target:
          $ref: "#/components/schemas/Target"
        collection:
          type: string
          description: Canonical collection id owning this file.
        path:
          type: string
          description: Relative file path within the collection, no leading slash.
        bytes: { type: integer }
        sha256:
          type: string
          description: SHA-256 hex digest of the file content.
        hot: { type: boolean }
        disc_coverage: { type: boolean }
    FilesResponse:
      type: object
      required: [target, page, per_page, total, pages, files]
      properties:
        target:
          $ref: "#/components/schemas/Target"
        page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        pages:
          type: integer
          minimum: 0
        files:
          type: array
          items:
            $ref: "#/components/schemas/FileRecord"
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }